Moved MetricFamiliesPerUser type to util package.
